+++ pycvodes/include/cvodes_cxx.hpp
@@ -22,8 +22,6 @@
 #include <cvodes/cvodes_impl.h> /* CVodeMem */
 #include <cvodes/cvodes_lapack.h>       /* prototype for CVDense */
 
-#include "anyode.hpp"
-
 
 namespace {
     class StreamFmt
@@ -51,17 +49,33 @@ namespace {
 
 namespace cvodes_cxx {
 
-    template<class T> void ignore( const T& ) { } // ignore compiler warnings about unused parameter
-
     //using sundials_cxx::nvector_serial::N_Vector; // native sundials vector
     using SVector = sundials_cxx::nvector_serial::Vector; // serial vector
 
     // Wrapper for Revision 4306 of cvodes.c
 
     enum class LMM : int {Adams=CV_ADAMS, BDF=CV_BDF}; // Linear multistep method
+    LMM lmm_from_name(std::string name){
+        if (name == "adams")
+            return LMM::Adams;
+        else if (name == "bdf")
+            return LMM::BDF;
+        else
+            throw std::runtime_error(StreamFmt() << "Unknown linear multistep method: " << name);
+    }
 
     enum class Task : int {Normal=CV_NORMAL, One_Step=CV_ONE_STEP};
-    enum class IterType : int {Functional=CV_FUNCTIONAL, Newton=CV_NEWTON};
+    enum class IterType : int {Functional=CV_FUNCTIONAL, Newton=CV_NEWTON, Undecided=-1};
+    IterType iter_type_from_name(std::string name){
+        if (name == "functional")
+            return IterType::Functional;
+        else if (name == "newton")
+            return IterType::Newton;
+        else if (name == "undecided")
+            return IterType::Undecided;
+        else
+            throw std::runtime_error(StreamFmt() << "Unknown iter_type: " << name);
+    }
     enum class IterLinSolEnum : int {GMRES=1, BICGSTAB=2, TFQMR=3};
     enum class PrecType : int {None=PREC_NONE, Left=PREC_LEFT,
             Right=PREC_RIGHT, Both=PREC_BOTH};
@@ -81,9 +95,15 @@ namespace cvodes_cxx {
     class CVodeIntegrator{ // Thin wrapper class of CVode in CVODES
     public:
         void *mem {nullptr};
-        std::size_t ny {0};
+        long int ny {0};  // cvodes uses a signed data type for this...
         CVodeIntegrator(const LMM lmm, const IterType iter) {
             this->mem = CVodeCreate(static_cast<int>(lmm), static_cast<int>(iter));
+            if (!this->mem)
+                throw std::bad_alloc(); // "CVodeCreate failed (allocation failed)."
+        }
+        ~CVodeIntegrator(){
+            if (this->mem)
+                CVodeFree(&(this->mem));
         }
         // init
         void init(CVRhsFn cb, realtype t0, N_Vector y) {
@@ -153,6 +173,9 @@ namespace cvodes_cxx {
         void set_max_num_steps(long int mxsteps){
             int status = CVodeSetMaxNumSteps(this->mem, mxsteps);
             check_flag(status);
+        }
+        long int get_max_num_steps(){
+            return static_cast<CVodeMem>(this->mem)->cv_mxstep;
         }
 
         // set_tol
@@ -448,9 +471,12 @@ namespace cvodes_cxx {
                                      CVodeGetReturnFlagName(flag));
         }
         std::pair<std::vector<realtype>, std::vector<realtype> >
-        adaptive(long int ny, const realtype x0, const realtype xend,
-                 const realtype * const y0, int nderiv, std::vector<int>& root_indices,
-                 bool return_on_root=false, long int mxsteps=0){
+        adaptive(const realtype x0,
+                 const realtype xend,
+                 const realtype * const y0,
+                 const unsigned nderiv,
+                 std::vector<int>& root_indices,
+                 bool return_on_root=false){
             std::vector<realtype> xout;
             std::vector<realtype> yout;
             realtype cur_t;
@@ -458,7 +484,8 @@ namespace cvodes_cxx {
             int idx = 0;
             SVector y {ny};
             SVector work {ny};
-            if (mxsteps == 0) { mxsteps = 500; } // cvodes default
+            long int mxsteps = get_max_num_steps();
+            if (mxsteps == 0) { mxsteps = 500; } // cvodes default (MXSTEP_DEFAULT)
             xout.push_back(x0);
             for (int i=0; i<ny; ++i){
                 y[i] = y0[i];
@@ -470,7 +497,7 @@ namespace cvodes_cxx {
                 for (int i=0; i<ny; ++i)
                     yout.push_back(work[i]);
             }
-            for (int di=1; di<nderiv; ++di){
+            for (unsigned di=1; di<nderiv; ++di){
                 for (int i=0; i<ny; ++i)  // higher order too expensive
                     yout.push_back(0);
             }
@@ -492,7 +519,7 @@ namespace cvodes_cxx {
                 for (int i=0; i<ny; ++i)
                     yout.push_back(y[i]);
                 // Derivatives for interpolation
-                for (int di=0; di<nderiv; ++di){
+                for (unsigned di=0; di<nderiv; ++di){
                     if (this->get_n_steps() < 2*(nderiv+1))
                         // Too few points collected
                         work.zero_out();
@@ -507,8 +534,13 @@ namespace cvodes_cxx {
             return std::pair<std::vector<realtype>, std::vector<realtype>>(xout, yout);
         }
 
-        void predefined(int nt, int ny, const realtype * const tout, const realtype * const y0,
-                        realtype * const yout, int nderiv, std::vector<int>& root_indices, std::vector<realtype>& root_out){
+        void predefined(const long int nt,  // sundials does not use unsigned types here...
+                        const realtype * const tout,
+                        const realtype * const y0,
+                        realtype * const yout,
+                        const unsigned nderiv,
+                        std::vector<int>& root_indices,
+                        std::vector<realtype>& root_out){
             realtype cur_t;
             int status;
             SVector y {ny};
@@ -522,7 +554,7 @@ namespace cvodes_cxx {
                 this->call_rhs(tout[0], y, work);
                 work.dump(&yout[ny]);
             }
-            for (int di=1; di<nderiv; ++di){
+            for (unsigned di=1; di<nderiv; ++di){
                 for (int i=0; i<ny; ++i)  // too expensive
                     yout[ny*(di+1) + i] = 0;
             }
@@ -543,7 +575,7 @@ namespace cvodes_cxx {
                 }
                 y.dump(&yout[ny*(iout*(nderiv+1))]);
                 // Derivatives for interpolation
-                for (int di=0; di<nderiv; ++di){
+                for (unsigned di=0; di<nderiv; ++di){
                     if (this->get_n_steps() < 2*(nderiv+1))
                         // Too few points collected
                         work.zero_out();
@@ -554,202 +586,18 @@ namespace cvodes_cxx {
             }
         }
 
-        ~CVodeIntegrator(){
-            if (this->mem)
-                CVodeFree(&(this->mem));
-        }
     };
 
-    int handle_status_(AnyODE::Status status){
-        switch (status){
-        case AnyODE::Status::success:
-            return 0;
-        case AnyODE::Status::recoverable_error:
-            return 1;
-        case AnyODE::Status::unrecoverable_error:
-            return -1;
-        default:
-            throw std::runtime_error("impossible (this is for silencing -Wreturn-type)");
-        }
-    }
-
-    template<class OdeSys>
-    int rhs_cb(realtype t, N_Vector y, N_Vector ydot, void *user_data){
-        auto& odesys = *static_cast<OdeSys*>(user_data);
-        AnyODE::Status status = odesys.rhs(t, NV_DATA_S(y), NV_DATA_S(ydot));
-        return handle_status_(status);
-    }
-
-    template<class OdeSys>
-    int roots_cb(realtype t, N_Vector y, realtype *gout, void *user_data){
-        auto& odesys = *static_cast<OdeSys*>(user_data);
-        AnyODE::Status status = odesys.roots(t, NV_DATA_S(y), gout);
-        if (status == AnyODE::Status::recoverable_error)
-            throw std::runtime_error("There are only unrecoverable errors for roots().");
-        return handle_status_(status);
-    }
-
-    template <class OdeSys>
-    int jac_dense_cb(long int N, realtype t,
-                     N_Vector y, N_Vector fy, DlsMat Jac, void *user_data,
-                     N_Vector tmp1, N_Vector tmp2, N_Vector tmp3){
-        // callback of req. signature wrapping OdeSys method.
-        ignore(N); ignore(tmp1); ignore(tmp2); ignore(tmp3);
-        auto& odesys = *static_cast<OdeSys*>(user_data);
-        AnyODE::Status status = odesys.dense_jac_cmaj(t, NV_DATA_S(y), NV_DATA_S(fy), DENSE_COL(Jac, 0), Jac->ldim);
-        return handle_status_(status);
-    }
-
-    template <typename OdeSys>
-    int jac_band_cb(long int N, long int mupper, long int mlower, realtype t,
-                    N_Vector y, N_Vector fy, DlsMat Jac, void *user_data,
-                    N_Vector tmp1, N_Vector tmp2, N_Vector tmp3){
-        ignore(N); ignore(tmp1); ignore(tmp2); ignore(tmp3);
-        auto& odesys = *static_cast<OdeSys*>(user_data);
-        if (odesys.get_mupper() != mupper)
-            throw std::runtime_error("mupper mismatch");
-        if (odesys.get_mlower() != mlower)
-            throw std::runtime_error("mlower mismatch");
-        AnyODE::Status status = odesys.banded_jac_cmaj(t, NV_DATA_S(y), NV_DATA_S(fy), Jac->data + Jac->s_mu - Jac->mu, Jac->ldim);
-        return handle_status_(status);
-    }
-
-
-    template <typename OdeSys>
-    int jac_times_vec_cb(N_Vector v, N_Vector Jv, realtype t, N_Vector y,
-                         N_Vector fy, void *user_data, N_Vector tmp){
-        // callback of req. signature wrapping OdeSys method.
-        ignore(tmp);
-        auto& odesys = *static_cast<OdeSys*>(user_data);
-        AnyODE::Status status = odesys.jac_times_vec(NV_DATA_S(v), NV_DATA_S(Jv), t, NV_DATA_S(y), NV_DATA_S(fy));
-        if (status == AnyODE::Status::recoverable_error)
-            throw std::runtime_error("There are only unrecoverable errors for JacTimesVec().");
-        return handle_status_(status);
-    }
-
-    template <typename OdeSys> // Section 4.6.9 Preconditioning in cvs_guide.pdf
-    int jac_prec_solve_cb(realtype t, N_Vector y, N_Vector fy, N_Vector r,
-                          N_Vector z, realtype gamma, realtype delta, int lr,
-                          void *user_data, N_Vector tmp){
-        // callback of req. signature wrapping OdeSys method.
-        ignore(tmp);  // delta used for iterative methods
-        double * ewt {nullptr};
-        auto& odesys = *static_cast<OdeSys*>(user_data);
-        // if (NV_LENGTH_S(tmp) >= odesys.get_ny()) {
-        //     ewt = NV_DATA_S(tmp)
-        // } else {
-        //     throw std::runtime_error("Implement allocating logic.");
-        // }
-        // odesys.integrator.get_err_weights(tmp); // use tmp to hold error weights
-        if (lr != 1)
-            throw std::runtime_error("Only left preconditioning implemented.");
-        AnyODE::Status status =  odesys.prec_solve_left(t, NV_DATA_S(y), NV_DATA_S(fy), NV_DATA_S(r),
-                                                        NV_DATA_S(z), gamma, delta, ewt);
-        return handle_status_(status);
-    }
-
-    template <typename OdeSys>
-    int prec_setup_cb(realtype t, N_Vector y, N_Vector fy, booleantype jok,
-                      booleantype *jcurPtr, realtype gamma, void *user_data,
-                      N_Vector tmp1, N_Vector tmp2, N_Vector tmp3){
-        // callback of req. signature wrapping OdeSys method.
-        ignore(tmp1); ignore(tmp2); ignore(tmp3);
-        auto& odesys = *static_cast<OdeSys*>(user_data);
-        bool jac_recomputed = false;
-        bool compute_jac = (jok == TRUE) ? false : true; // TRUE, FALSE sundials macros
-        AnyODE::Status status = odesys.prec_setup(t, NV_DATA_S(y), NV_DATA_S(fy), compute_jac, jac_recomputed, gamma);
-        (*jcurPtr) = (jac_recomputed) ? TRUE : FALSE;
-        return handle_status_(status);
-    }
-
-    template <class OdeSys>
-    CVodeIntegrator get_integrator(OdeSys * odesys,
-                                   const std::vector<realtype> atol,
-                                   const realtype rtol,
-                                   const int lmm,  // Cython does not supported enum class...
-                                   const realtype * const y0,
-                                   const realtype t0,
-                                   const realtype dx0=0.0,
-                                   const realtype dx_min=0.0,
-                                   const realtype dx_max=0.0,
-                                   const long int mxsteps=0,
-                                   const bool with_jacobian=false,
-                                   const int iter_type=0,
-                                   const int linear_solver=0,
-                                   const int maxl=0,
-                                   const realtype eps_lin=0.0
-                                   )
-    {
-        const int ny = odesys->get_ny();
-        CVodeIntegrator integr {(lmm == CV_BDF) ? LMM::BDF : LMM::Adams,
-                (iter_type == 1) ? IterType::Functional : IterType::Newton};
-        integr.set_user_data(static_cast<void *>(odesys));
-        integr.init(rhs_cb<OdeSys>, t0, y0, ny);
-        integr.root_init(odesys->get_nroots(), roots_cb<OdeSys>);
-        if (atol.size() == 1){
-            integr.set_tol(rtol, atol[0]);
-        }else{
-            integr.set_tol(rtol, atol);
-        }
-        integr.set_init_step(dx0);
-        if (dx_min != 0.0)
-            integr.set_min_step(dx0);
-        if (dx_max != 0.0)
-            integr.set_max_step(dx0);
-        if (mxsteps)
-            integr.set_max_num_steps(mxsteps);
-
-
-        if (iter_type == 2){
-            // Newton iteration --> we need a linear solver:
-            switch(linear_solver){
-            case 1:
-                integr.set_linear_solver_to_dense(ny);
-                if (with_jacobian)
-                    integr.set_dense_jac_fn(jac_dense_cb<OdeSys >);
-                break;
-            case 2:
-                integr.set_linear_solver_to_banded(ny, odesys->get_mupper(), odesys->get_mlower());
-                if (with_jacobian)
-                    integr.set_band_jac_fn(jac_band_cb<OdeSys >);
-                break;
-            case 10:
-            case 11:
-                integr.set_linear_solver_to_iterative(IterLinSolEnum::GMRES, maxl); break;
-            case 20:
-                integr.set_linear_solver_to_iterative(IterLinSolEnum::BICGSTAB, maxl); break;
-            case 30:
-                integr.set_linear_solver_to_iterative(IterLinSolEnum::TFQMR, maxl); break;
-            default:
-                throw std::runtime_error("Invalid linear_solver");
-            }
-            if (linear_solver >= 10){
-                if (!with_jacobian)
-                    throw std::runtime_error("Iterative method requires an (approximate) jacobian");
-                integr.set_prec_type(PrecType::Left);
-                integr.set_iter_eps_lin(eps_lin);
-                integr.set_jac_times_vec_fn(jac_times_vec_cb<OdeSys>);
-                integr.set_preconditioner(prec_setup_cb<OdeSys>, jac_prec_solve_cb<OdeSys>);
-                if (linear_solver == 10 || linear_solver == 11) // GMRES
-                    integr.set_gram_schmidt_type((linear_solver == 10) ? GramSchmidtType::Modified : GramSchmidtType::Classical);
-                else if (linear_solver == 20 or linear_solver == 30) // BiCGStab, TFQMR
-                    ;
-                else
-                    throw std::runtime_error("Unknown linear_solver.");
-            }
-        }
-        return integr;
-    }
-
-    void set_integration_info(std::unordered_map<std::string, int>& info, const CVodeIntegrator& integrator,
-                              int iter_type, int linear_solver){
+    void set_integration_info(std::unordered_map<std::string, int>& info,
+                              const CVodeIntegrator& integrator,
+                              IterType iter_type, int linear_solver){
         info["n_steps"] = integrator.get_n_steps();
         info["n_rhs_evals"] = integrator.get_n_rhs_evals();
         info["n_lin_solv_setups"] = integrator.get_n_lin_solv_setups();
         info["n_err_test_fails"] = integrator.get_n_err_test_fails();
         info["n_nonlin_solv_iters"] = integrator.get_n_nonlin_solv_iters();
         info["n_nonlin_solv_conv_fails"] = integrator.get_n_nonlin_solv_conv_fails();
-        if (iter_type == 2){
+        if (iter_type == IterType::Newton){
             if (linear_solver >= 10) {  // iterative linear solver
                 info["krylov_n_lin_iters"] = integrator.get_n_lin_iters();
                 info["krylov_n_prec_evals"] = integrator.get_n_prec_evals();
@@ -762,102 +610,6 @@ namespace cvodes_cxx {
                 info["dense_n_dls_rhs_evals"] = integrator.get_n_dls_rhs_evals();
             }
         }
-    }
-
-    template <class OdeSys>
-    std::pair<std::vector<realtype>, std::vector<realtype> >
-    simple_adaptive(OdeSys * const odesys,
-                    const std::vector<realtype> atol,
-                    const realtype rtol,
-                    const int lmm,
-                    const realtype * const y0,
-                    const realtype t0,
-                    const realtype tend,
-                    std::vector<int>& root_indices,
-                    const realtype dx0=0.0,
-                    const realtype dx_min=0.0,
-                    const realtype dx_max=0.0,
-                    const long int mxsteps=0,
-                    const bool with_jacobian=false,
-                    int iter_type=0,
-                    int linear_solver=0,
-                    const int maxl=0,
-                    const realtype eps_lin=0.0,
-                    const int nderiv=0,
-                    bool return_on_root=false
-                    ){
-        // iter_type == 0 => 1 if lmm == CV_ADAMS else 2
-        // iter_type == 1 => Functional (ignore linear_solver)
-        // iter_type == 2 => Newton
-
-        // linear_solver ==  0 => 1 if get_mlower() == -1 else 2
-        // linear_solver ==  1 => Direct (dense LU-factorization)
-        // linear_solver ==  2 => Direct (banded LU-factorization)
-        // linear_solver == 10 => Iterative, GMRES, Modified Gram-Schmidt
-        // linear_solver == 11 => Iterative, GMRES, Classical Gram-Schmidt
-        // linear_solver == 20 => Iterative, Bi-CGStab (maxl => maximum dimension of Krylov subspace)
-        // linear_solver == 30 => Iterative, TFQMR (maxl => maximum dimension of Krylov subspace)
-        if (iter_type == 0)
-            iter_type = (lmm == CV_ADAMS) ? 1 : 2;
-        if (iter_type != 1 and iter_type != 2)
-            throw std::runtime_error("Invalid iter_type");
-        if (linear_solver == 0)
-            linear_solver = (odesys->get_mlower() == -1) ? 1 : 2;
-
-        auto integr = get_integrator<OdeSys>(odesys, atol, rtol, lmm, y0, t0, dx0, dx_min, dx_max, mxsteps,
-                                             with_jacobian, iter_type, linear_solver, maxl, eps_lin);
-        odesys->integrator = static_cast<void*>(&integr);
-        auto result = integr.adaptive(odesys->get_ny(), t0, tend, y0, nderiv, root_indices, return_on_root, mxsteps);
-        odesys->last_integration_info.clear();
-        set_integration_info(odesys->last_integration_info, integr, iter_type, linear_solver);
-        return result;
-    }
-
-    template <class OdeSys>
-    void simple_predefined(OdeSys * const odesys,
-                           const std::vector<realtype> atol,
-                           const realtype rtol,
-                           const int lmm,
-                           const realtype * const y0,
-                           const std::size_t nout,
-                           const realtype * const tout,
-                           realtype * const yout,
-                           std::vector<int>& root_indices,
-                           std::vector<double>& root_out,
-                           const realtype dx0=0.0,
-                           const realtype dx_min=0.0,
-                           const realtype dx_max=0.0,
-                           const long int mxsteps=0,
-                           const bool with_jacobian=false,
-                           int iter_type=0,
-                           int linear_solver=0,
-                           const int maxl=0,
-                           const realtype eps_lin=0.0,
-                           const int nderiv=0
-                           ){
-        // iter_type == 0 => 1 if lmm == CV_ADAMS else 2
-        // iter_type == 1 => Functional (ignore linear_solver)
-        // iter_type == 2 => Newton
-
-        // linear_solver ==  0 => 1 if get_mlower() == -1 else 2
-        // linear_solver ==  1 => Direct (dense LU-factorization)
-        // linear_solver ==  2 => Direct (banded LU-factorization)
-        // linear_solver == 10 => Iterative, GMRES, Modified Gram-Schmidt
-        // linear_solver == 11 => Iterative, GMRES, Classical Gram-Schmidt
-        // linear_solver == 20 => Iterative, Bi-CGStab (maxl => maximum dimension of Krylov subspace)
-        // linear_solver == 30 => Iterative, TFQMR (maxl => maximum dimension of Krylov subspace)
-        if (iter_type == 0)
-            iter_type = (lmm == CV_ADAMS) ? 1 : 2;
-        if (iter_type != 1 and iter_type != 2)
-            throw std::runtime_error("Invalid iter_type");
-        if (linear_solver == 0)
-            linear_solver = (odesys->get_mlower() == -1) ? 1 : 2;
-        auto integr = get_integrator<OdeSys>(odesys, atol, rtol, lmm, y0, tout[0], dx0, dx_min, dx_max, mxsteps,
-                                             with_jacobian, iter_type, linear_solver, maxl, eps_lin);
-        odesys->integrator = static_cast<void*>(&integr);
-        integr.predefined(nout, odesys->get_ny(), tout, y0, yout, nderiv, root_indices, root_out);
-        odesys->last_integration_info.clear();
-        set_integration_info(odesys->last_integration_info, integr, iter_type, linear_solver);
     }
 
 }