+++ tests/testing_utils.hpp
@@ -1,20 +1,15 @@
 //#include <unordered_map>
 #include "cvodes_cxx.hpp"
-#include "anyode"
-
-int rhs_cb(double t, N_Vector y, N_Vector f, void * user_data){
-    cvodes_cxx::ignore(t); cvodes_cxx::ignore(user_data);
-    NV_DATA_S(f)[0] = -NV_DATA_S(y)[0];
-    return 0;
-}
+#include "anyode/anyode.hpp"
 
 struct Decay : public AnyODE::OdeSysBase {
     double m_k;
 
     Decay(double k) : m_k(k) {}
     int get_ny() const override { return 1; }
-    void rhs(double t, const double * const __restrict__ y, double * const __restrict__ f) override {
-        cvodes_cxx::ignore(t);
+    AnyODE::Status rhs(double t, const double * const __restrict__ y, double * const __restrict__ f) override {
+        AnyODE::ignore(t);
         f[0] = -y[0];
+        return AnyODE::Status::success;
     }
 };